fix(media_control): do not persist volume when is initial config
Prevent the initial volume value from being persisted in the localStorage.
Possible issues: When a player instance is created with mute option enabled, the value 0 is persisted. When a new instance is created without this option (or mute=false), the 0 value will be restored as initial volume, causing an inconsistency.